+++ src/main/java/niv/flowstone/config/ConfigurationLoader.java
@@ -1,5 +1,6 @@
 package niv.flowstone.config;
 
+import static niv.flowstone.Flowstone.LOGGER;
 import static niv.flowstone.Flowstone.MOD_ID;
 
 import java.io.File;
@@ -21,6 +22,7 @@ public final class ConfigurationLoader {
     private static final long DELAY = 5000; // ms (5s)
 
     private static final Gson gson = new GsonBuilder()
+            .setLenient()
             .setPrettyPrinting()
             .create();
 
@@ -54,8 +56,12 @@ public final class ConfigurationLoader {
         var now = System.currentTimeMillis();
         if (timestamp + DELAY < now) {
             var file = configurationFile.get();
-            if (create(file) || read(file)) {
-                write(file);
+            try {
+                if (create(file) || read(file)) {
+                    write(file);
+                }
+            } catch (IllegalStateException ex) {
+                LOGGER.warn(ex.getMessage(), ex);
             }
             timestamp = now;
         }
@@ -63,7 +69,9 @@ public final class ConfigurationLoader {
 
     private static final boolean create(File file) {
         try {
-            return (file.getParentFile().isDirectory() || file.getParentFile().mkdirs()) && file.createNewFile();
+            return !file.isFile()
+                    && (file.getParentFile().isDirectory() || file.getParentFile().mkdirs())
+                    && file.createNewFile();
         } catch (IOException ex) {
             throw new IllegalStateException("Failed to create configuration file", ex);
         }